Added charts

+++ lib/repos/recipe_statistics/recipe_statistics_repo.dart
@@ -15,7 +15,7 @@ abstract class RecipeStatisticsRepo {
     required DateTime endDate,
   });
 
-  Future<Map<String, int>> getGroceryAmountBetween({
+  Future<Map<String, double>> getGroceryAmountBetween({
     required DateTime startDate,
     required DateTime endDate,
   });
